fix: allow picking cwd in multi-root workspaces
Currently, the VS Code API doesn't allow determining from which root a task
comes from. To allow tasks to be run within the correct working directory
in multi-root workspaces, we show a pick dialog to the user with a list of
all the current roots. Better support for multi-root workspaces will be
provided in a future change, once microsoft/vscode#137552 is resolved.

Resolves #19.