fix: prepend sudo to austin command on MacOS
Austin requires the use of sudo on MacOS. This means that the
"Profile with Austin" command did not work because it was being run
without sudo. This change ensures that Austin is run with sudo on
MacOS, thus removing this limitation.

Note that this still requires that the austin binary is listed in
the sudoers file to run without password.

Addresses #35.