fix: ensure line end is at least line (#60)
A bug in the MOJO generation for Python <= 3.11 from Austin caused the
heat map to be wrongly visualised in the extension. That is due to
Austin setting the line end to the first line of the function when it
should be 0. We make sure that the line end is always at least line
when trying to visualise heat maps and line stats in the editor to add
support for the versions of Austin with the bug.